requirements.txt
# 智知因 - Python依赖清单
# 生产级依赖管理

# === 核心框架 ===
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.30.0
python-multipart==0.0.6

# === 数据库 ===
# SQLite (内置)
chromadb==0.4.22
langchain==0.1.10
langchain-community==0.0.25
langchain-core==0.1.30

# === 大模型 & Agent ===
langgraph==0.0.35
langchain-deepseek==0.1.0
langchain-openai==0.0.8

# === Embedding ===
sentence-transformers==2.3.1

# === PDF处理 ===
pypdf==4.0.1

# === 工具库 ===
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
loguru==0.7.2
sse-starlette==2.0.0
psutil==5.9.8

# === 测试 ===
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0
